working base_pytest2
